pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = ""
authors = ["Ivan Feofanov <feofanov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django-allauth = "^65.2.0"
black = "^24.10.0"
requests = "^2.32.3"
pyjwt = "^2.10.0"
uvicorn = "^0.32.1"
cryptography = "^43.0.3"
dj-database-url = "^2.3.0"
psycopg2-binary = "^2.9.10"
django-formify = "^0.0.7"
django-widget-tweaks = "^1.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
